+++ src/main/java/net/soulsweaponry/items/crossbow/KrakenSlayerCrossbow.java
@@ -20,7 +20,7 @@ public class KrakenSlayerCrossbow extends ModdedCrossbow {
     public KrakenSlayerCrossbow(Settings settings, Supplier<Ingredient> repairIngredientSupplier) {
         super(settings, repairIngredientSupplier);
         this.addTooltipAbility(TooltipAbilities.FAST_PULL, TooltipAbilities.THIRD_SHOT);
-        this.configure(new RangedConfig(ConfigConstructor.kraken_slayer_crossbow_pull_time_ticks, ConfigConstructor.kraken_slayer_crossbow_damage, ConfigConstructor.kraken_slayer_crossbow_max_velocity));
+        this.configure(new RangedConfig((int) ConfigConstructor.kraken_slayer_crossbow_pull_time_ticks, ConfigConstructor.kraken_slayer_crossbow_damage, ConfigConstructor.kraken_slayer_crossbow_max_velocity));
     }
 
     @Override
